feat(native-app): extract to a function and rename index to home
