+++ apps/native/app/src/app/_layout.tsx
@@ -1,11 +1,10 @@
 import { ThemeProvider as AppThemeProvider } from '@/components/providers/theme-provider'
 import { PromptModal } from '@/components/prompt-modal'
 import { ToastHost } from '@/components/toast'
-import { Stack, useSegments } from 'expo-router'
+import { Stack } from 'expo-router'
 import * as SplashScreen from 'expo-splash-screen'
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useState } from 'react'
 import 'react-native-reanimated'
-import { DdRum } from '@datadog/mobile-react-native'
 
 import { LocaleProvider } from '../components/providers/locale-provider'
 import { OfflineProvider } from '../components/providers/offline-provider'
@@ -36,6 +35,7 @@ import { IBMPlexSans_500Medium_Italic } from '@expo-google-fonts/ibm-plex-sans/5
 import { IBMPlexSans_600SemiBold_Italic } from '@expo-google-fonts/ibm-plex-sans/600SemiBold_Italic'
 import { IBMPlexSans_700Bold_Italic } from '@expo-google-fonts/ibm-plex-sans/700Bold_Italic'
 import { setupEventHandlers } from '../utils/lifecycle/setup-event-handlers'
+import { useDatadogViewTracking } from '../hooks/use-datadog-view-tracking'
 
 export { ErrorBoundary } from 'expo-router'
 
@@ -185,25 +185,7 @@ function RootLayoutNav({
 }: {
   apolloClient: ApolloClient<NormalizedCacheObject>
 }) {
-  // Datadog RUM view tracking. Uses route segment templates (e.g. inbox/[id])
-  // rather than the resolved pathname so dynamic params like document IDs
-  // don't leak into Datadog.
-  const segments = useSegments()
-  const viewName =
-    segments.filter((s) => !s.startsWith('(')).join('/') || 'index'
-  const previousViewRef = useRef<string | null>(null)
-
-  useEffect(() => {
-    if (__DEV__) {
-      return
-    }
-    const previous = previousViewRef.current
-    if (previous && previous !== viewName) {
-      DdRum.stopView(previous).catch(() => {})
-    }
-    DdRum.startView(viewName, viewName).catch(() => {})
-    previousViewRef.current = viewName
-  }, [viewName])
+  useDatadogViewTracking()
 
   return (
     <LocaleProvider>
